+++ src/rideables/NVMGraph.hpp
@@ -23,7 +23,7 @@
 /**
  * SimpleGraph class.  Labels are of templated type K.
  */
-template <size_t numVertices = 1024, size_t meanEdgesPerVertex=20, size_t vertexLoad = 50>
+template <int numVertices = 1024, int meanEdgesPerVertex=20, int vertexLoad = 50>
 class NVMGraph : public RGraph {
 
     public:
@@ -317,7 +317,7 @@ class NVMGraph : public RGraph {
             bool retval = true;
             // Randomly sample vertices...
             std::vector<int> vec;
-            for (size_t i = 0; i < meanEdgesPerVertex * 100 / vertexLoad; i++) {
+            for (int i = 0; i < meanEdgesPerVertex * 100 / vertexLoad; i++) {
                 int u = uniformVertex(vertexGen);
                 if (u == i) {
                     continue;
